Wow, so everything worked after last committ (I think) but performance was super slow. C# took 2 seconds and my Java version took 45 seconds. Unacceptable. After using the awesome profiler in Netbeans 17 I found that the problem was ONLY the PriorityQueue.remove function. Because it was doing a linear search every time.
So I went ahead an implemented my own PriorityQueue that uses binary search for the remove.
But then it was slow on the add operation because here I was doing a linear search for the insert post myself.
So I changed the add to be binary searching as well and now I also have 2 seconds!!!! Yay!
Also I used LinkedList first, but that was also slow, because the get in my binary search was doing a linear linked list crawl.

I added a better hashCode method to Pair
I made the compareTo return 0
Also I did some refactorings and other code cleanups.